requirements.txt
pillow>=9.0.0
numpy>=1.20.0
PyQt6>=6.2.0
opencv-python>=4.5.0
